+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Note that since we don't clearly distinguish between a public and private interf
 
 - Add `inverted` option to `xrayShaded` parameter
 - Model-export extension: Add ability to set a file name for structures
+- Add `contextHash` to `SizeTheme`
 
 ## [v3.36.1] - 2023-06-11